+++ changelog.md
@@ -8,11 +8,14 @@
 - Added a new object to handle all plugin settings
 - Added support for any post types which enabled title and editor as we need them for the simplification of texts
 - Added support to simply taxonomy terms
+- Added support for easy and plain read in english
+- Added support for FALC in french
 - Added support for the PageBuilders Beaver Builder, Bold Page Builder, Brizy, Kubio, SiteOrigin, SeedProd and Visual Composer
 - Added option to reset the plugin in backend settings (in preparation for Cyber Resilience Act)
 - Added Taskfile as alternative release building method for the plugin
 - Added check against WordPress Plugin Checker during each release
 - Added info in footer of admin where our plugin adds options or show settings
+- Added support to simplify the tests in plugin "Popup Maker"
 
 ### Changed